+++ src/Markdown.MAML/Transformer/ModelTransformerVersion2.cs
@@ -4,9 +4,6 @@ using Markdown.MAML.Resources;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Collections.Specialized;
 using Markdown.MAML.Parser;
 
 namespace Markdown.MAML.Transformer
@@ -17,6 +14,8 @@ namespace Markdown.MAML.Transformer
             new List<Tuple<string, Dictionary<string, MamlParameter>>>();
 
         public static readonly string ALL_PARAM_SETS_MONIKER = "(All)";
+
+        private string DefaultParameterSet;
 
         public ModelTransformerVersion2() : this(null) { }
 
@@ -101,8 +100,10 @@ namespace Markdown.MAML.Transformer
             MamlSyntax syntax;
             while ((syntax = SyntaxEntryRule()) != null)
             {
-                // We ignore Syntax section from markdown in the transformation
-                // We get all nessesary info from parameters section
+                //this is the only way to retain information on which syntax is the default 
+                // without adding new members to command object.
+                //Though the cmdlet object, does have a member which contains the default syntax name only.
+                if (syntax.IsDefault) { commmand.Syntax.Add(syntax); }
             }
         }
 
@@ -122,7 +123,8 @@ namespace Markdown.MAML.Transformer
                 // if header is omitted
                 syntax = new MamlSyntax()
                 {
-                    ParameterSetName = "__AllParameterSets"
+                    ParameterSetName = ALL_PARAM_SETS_MONIKER,
+                    IsDefault = true
                 };
             }
             else
@@ -133,9 +135,11 @@ namespace Markdown.MAML.Transformer
                     return null;
                 }
 
+                bool isDefault = headingNode.Text.EndsWith(MarkdownStrings.DefaultParameterSetModifier);
                 syntax = new MamlSyntax()
                 {
-                    ParameterSetName = headingNode.Text
+                    ParameterSetName = isDefault ? headingNode.Text.Substring(0, headingNode.Text.Length - MarkdownStrings.DefaultParameterSetModifier.Length) : headingNode.Text,
+                    IsDefault = isDefault
                 };
 
                 var codeBlock = CodeBlockRule();
@@ -186,7 +190,15 @@ namespace Markdown.MAML.Transformer
         private void GatherSyntax(MamlCommand command)
         {
             var parameterSetNames = GetParameterSetNames();
+            var defaultSetName = string.Empty;
 
+            if(command.Syntax.Count == 1 && command.Syntax[0].IsDefault)
+            {
+                //checks for existing IsDefault paramset and remove it while saving the name
+                defaultSetName = command.Syntax[0].ParameterSetName;
+                command.Syntax.Remove(command.Syntax[0]);
+            }
+            
             if (parameterSetNames.Count == 0)
             {
                 // special case: there are no parameters and hence there is only one parameter set
@@ -201,7 +213,7 @@ namespace Markdown.MAML.Transformer
                 {
                     if (parameterSetNames.Count == 1)
                     {
-                        // special case: there is only one parameter set and it's the deafult one
+                        // special case: there is only one parameter set and it's the default one
                         // we don't specify the name in this case.
                     }
                     else
@@ -211,7 +223,9 @@ namespace Markdown.MAML.Transformer
                 }
                 else
                 {
-                    syntax.ParameterSetName = setName;
+                    syntax.ParameterSetName = StringComparer.OrdinalIgnoreCase.Equals(syntax.ParameterSetName, defaultSetName) 
+                        ? string.Format("{0}{1}", setName, MarkdownStrings.DefaultParameterSetModifier) 
+                        : setName;
                 }
 
                 FillUpSyntax(syntax, setName);