+++ test/Markdown.MAML.Test/Renderer/MarkdownV2RendererTests.cs
@@ -22,9 +22,9 @@ namespace Markdown.MAML.Test.Renderer
             Assert.Equal(@"\\\\\\\<", MarkdownV2Renderer.GetEscapedMarkdownText(@"\\\<"));
             Assert.Equal(@"\", MarkdownV2Renderer.GetEscapedMarkdownText(@"\"));
             Assert.Equal(@"\\\\", MarkdownV2Renderer.GetEscapedMarkdownText(@"\\"));
-            Assert.Equal(@"\\\(", MarkdownV2Renderer.GetEscapedMarkdownText(@"\("));
-            Assert.Equal(@"\(", MarkdownV2Renderer.GetEscapedMarkdownText(@"("));
-            Assert.Equal(@"\)", MarkdownV2Renderer.GetEscapedMarkdownText(@")"));
+            Assert.Equal(@"\\(", MarkdownV2Renderer.GetEscapedMarkdownText(@"\("));
+            Assert.Equal(@"(", MarkdownV2Renderer.GetEscapedMarkdownText(@"("));
+            Assert.Equal(@")", MarkdownV2Renderer.GetEscapedMarkdownText(@")"));
             Assert.Equal(@"\[", MarkdownV2Renderer.GetEscapedMarkdownText(@"["));
             Assert.Equal(@"\]", MarkdownV2Renderer.GetEscapedMarkdownText(@"]"));
             Assert.Equal(@"\`", MarkdownV2Renderer.GetEscapedMarkdownText(@"`"));
@@ -63,12 +63,54 @@ Workflow [<WorkflowCommonParameters>] [<CommonParameters>]
 ## PARAMETERS
 
 ### WorkflowCommonParameters
-This cmdlet supports the following workflow common parameters: -PSParameterCollection, -PSComputerName, -PSCredential, -PSConnectionRetryCount, -PSConnectionRetryIntervalSec, -PSRunningTimeoutSec, -PSElapsedTimeoutSec, -PSPersist, -PSAuthentication, -PSAuthenticationLevel, -PSApplicationName, -PSPort, -PSUseSSL, -PSConfigurationName, -PSConnectionURI, -PSAllowRedirection, -PSSessionOption, -PSCertificateThumbprint, -PSPrivateMetadata, -AsJob, -JobName, and –InputObject.
-For more information, see about_WorkflowCommonParameters \(http://go.microsoft.com/fwlink/p/?LinkID=533952\).
+This cmdlet supports the following workflow common parameters: -PSParameterCollection, -PSComputerName, -PSCredential, -PSConnectionRetryCount, -PSConnectionRetryIntervalSec, -PSRunningTimeoutSec, -PSElapsedTimeoutSec, -PSPersist, -PSAuthentication, -PSAuthenticationLevel, -PSApplicationName, -PSPort, -PSUseSSL, -PSConfigurationName, -PSConnectionURI, -PSAllowRedirection, -PSSessionOption, -PSCertificateThumbprint, -PSPrivateMetadata, -AsJob, -JobName, and -InputObject.
+For more information, see about_WorkflowCommonParameters (http://go.microsoft.com/fwlink/p/?LinkID=533952).
 
 ### CommonParameters
 This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable, -InformationAction, -InformationVariable, -OutVariable, -OutBuffer, -PipelineVariable, -Verbose, -WarningAction, and -WarningVariable.
-For more information, see about_CommonParameters \(http://go.microsoft.com/fwlink/?LinkID=113216\).
+For more information, see about_CommonParameters (http://go.microsoft.com/fwlink/?LinkID=113216).
+
+## INPUTS
+
+## OUTPUTS
+
+## NOTES
+
+## RELATED LINKS
+
+", markdown);
+        }
+
+        [Fact]
+        public void RendererNormalizeQuotesAndDashes()
+        {
+            var renderer = new MarkdownV2Renderer(ParserMode.Full);
+            MamlCommand command = new MamlCommand()
+            {
+                Name = "Test-Quotes",
+                Description = @"”“‘’––-"
+            };
+
+            string markdown = renderer.MamlModelToString(command, null);
+            Assert.Equal(@"---
+schema: 2.0.0
+---
+
+# Test-Quotes
+## SYNOPSIS
+
+## SYNTAX
+
+## DESCRIPTION
+""""''---
+
+## EXAMPLES
+
+## PARAMETERS
+
+### CommonParameters
+This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable, -InformationAction, -InformationVariable, -OutVariable, -OutBuffer, -PipelineVariable, -Verbose, -WarningAction, and -WarningVariable.
+For more information, see about_CommonParameters (http://go.microsoft.com/fwlink/?LinkID=113216).
 
 ## INPUTS
 
@@ -212,7 +254,7 @@ Accept wildcard characters: True
 
 ### CommonParameters
 This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable, -InformationAction, -InformationVariable, -OutVariable, -OutBuffer, -PipelineVariable, -Verbose, -WarningAction, and -WarningVariable.
-For more information, see about_CommonParameters \(http://go.microsoft.com/fwlink/?LinkID=113216\).
+For more information, see about_CommonParameters (http://go.microsoft.com/fwlink/?LinkID=113216).
 
 ## INPUTS
 
@@ -282,7 +324,8 @@ Second line.
 
             var syntax2 = new MamlSyntax()
             {
-                ParameterSetName = "SecondSyntax"
+                ParameterSetName = "SecondSyntax",
+                IsDefault = true
             };
 
             syntax2.Parameters.Add(commonParam);
@@ -306,7 +349,7 @@ schema: 2.0.0
 Get-Foo -Common <String> -First <String> [<CommonParameters>]
 ```
 
-### SecondSyntax
+### SecondSyntax (Default)
 ```
 Get-Foo -Common <String> -Second <String> [<CommonParameters>]
 ```
@@ -358,7 +401,7 @@ Accept wildcard characters: False
 
 ### CommonParameters
 This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable, -InformationAction, -InformationVariable, -OutVariable, -OutBuffer, -PipelineVariable, -Verbose, -WarningAction, and -WarningVariable.
-For more information, see about_CommonParameters \(http://go.microsoft.com/fwlink/?LinkID=113216\).
+For more information, see about_CommonParameters (http://go.microsoft.com/fwlink/?LinkID=113216).
 
 ## INPUTS